Fix dependencies
