+++ crates/burn-backend/src/backend/device.rs
@@ -1,8 +1,7 @@
 pub use burn_std::device::*;
+pub use burn_std::{DeviceError, DeviceSettings};
 use burn_std::{BoolDType, BoolStore, DType, FloatDType, IntDType};
 
-use alloc::format;
-use alloc::string::String;
 use burn_std::stub::RwLock;
 
 #[cfg(target_has_atomic = "ptr")]
@@ -10,7 +9,6 @@ use alloc::sync::Arc;
 
 #[cfg(not(target_has_atomic = "ptr"))]
 use portable_atomic_util::Arc;
-use thiserror::Error;
 
 use core::any::TypeId;
 
@@ -31,39 +29,6 @@ pub trait DeviceOps: Clone + Default + PartialEq + Send + Sync + core::fmt::Debu
     /// Returns the [device id](DeviceId).
     fn id(&self) -> DeviceId {
         self.to_id()
-    }
-}
-
-/// Settings controlling the default data types for a specific device.
-///
-/// These settings are managed in a global registry that enforces strict initialization semantics:
-///
-/// 1. Manual Initialization: You can set these once at the start of your program using [`set_default_dtypes`].
-/// 2. Default Initialization: If an operation (like creating a tensor) occurs before manual initialization,
-///    the settings are permanently locked to their default values.
-/// 3. Immutability: Once initialized, settings cannot be changed. This ensures consistent behavior across
-///    all threads and operations.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct DeviceSettings {
-    /// Default floating-point data type.
-    pub float_dtype: FloatDType,
-    /// Default integer data type.
-    pub int_dtype: IntDType,
-    /// Default bool data type.
-    pub bool_dtype: BoolDType,
-}
-
-impl DeviceSettings {
-    fn new(
-        float_dtype: impl Into<FloatDType>,
-        int_dtype: impl Into<IntDType>,
-        bool_dtype: impl Into<BoolDType>,
-    ) -> Self {
-        Self {
-            float_dtype: float_dtype.into(),
-            int_dtype: int_dtype.into(),
-            bool_dtype: bool_dtype.into(),
-        }
     }
 }
 
@@ -207,46 +172,6 @@ fn default_float<B: Backend>() -> FloatDType {
 
 fn default_int<B: Backend>() -> IntDType {
     <B::IntElem as crate::Element>::dtype().into()
-}
-
-/// Errors that can occur during device-related operations.
-///
-/// This covers errors related to hardware capability mismatches, such as
-/// requesting a data type not supported by the device, and configuration
-/// errors like attempting to change a settings in an invalid context.
-#[derive(Debug, Error)]
-pub enum DeviceError {
-    /// Unsupported data type by the device.
-    #[error("Device {device} does not support the requested data type {dtype:?}")]
-    UnsupportedDType {
-        /// The string representation of the device.
-        device: String,
-        /// The data type that caused the error.
-        dtype: DType,
-    },
-    /// Device settings have already been initialized.
-    #[error("Device {device} settings have already been initialized")]
-    AlreadyInitialized {
-        /// The string representation of the device.
-        device: String,
-    },
-}
-
-impl DeviceError {
-    /// Helper to create a [`DeviceError::UnsupportedDType`] from any device.
-    pub fn unsupported_dtype<D: DeviceOps>(device: &D, dtype: DType) -> Self {
-        Self::UnsupportedDType {
-            device: format!("{device:?}"),
-            dtype,
-        }
-    }
-
-    /// Helper to create a [`DeviceError::AlreadyInitialized`] from any device.
-    pub fn already_initialized<D: DeviceOps>(device: &D) -> Self {
-        Self::AlreadyInitialized {
-            device: format!("{device:?}"),
-        }
-    }
 }
 
 fn check_dtype_support<B: Backend>(
@@ -447,15 +372,12 @@ mod tests {
 
     impl DeviceOps for TestDeviceB {}
 
-    // Test defaults
-    impl DeviceSettings {
-        fn defaults() -> Self {
-            DeviceSettings::new(FloatDType::F32, IntDType::I32, BoolDType::Native)
-        }
+    fn default_settings() -> DeviceSettings {
+        DeviceSettings::new(FloatDType::F32, IntDType::I32, BoolDType::Native)
     }
 
     fn get_test_device_settings<D: DeviceOps>(device: &D) -> DeviceSettings {
-        DeviceSettingsRegistry::get_or_insert(device, DeviceSettings::defaults)
+        DeviceSettingsRegistry::get_or_insert(device, default_settings)
     }
 
     #[test]
@@ -469,7 +391,7 @@ mod tests {
         let s2 = get_test_device_settings(&device);
 
         assert_eq!(s1, s2);
-        assert_eq!(s1, DeviceSettings::defaults());
+        assert_eq!(s1, default_settings());
     }
 
     #[test]
@@ -505,7 +427,7 @@ mod tests {
 
         assert_ne!(s1, s2);
         assert_eq!(s1, settings);
-        assert_eq!(s2, DeviceSettings::defaults());
+        assert_eq!(s2, default_settings());
     }
 
     #[test]
@@ -523,7 +445,7 @@ mod tests {
         let s2 = get_test_device_settings(&d2);
 
         assert_ne!(s1, s2);
-        assert_eq!(s1, DeviceSettings::defaults());
+        assert_eq!(s1, default_settings());
         assert_eq!(s2, settings);
     }
 
@@ -554,7 +476,7 @@ mod tests {
         let settings = DeviceSettings::new(FloatDType::F16, IntDType::I32, BoolDType::Native);
         initialize_unchecked(&device, settings).unwrap();
 
-        let result = initialize_unchecked(&device, DeviceSettings::defaults());
+        let result = initialize_unchecked(&device, default_settings());
         assert!(matches!(
             result,
             Err(DeviceError::AlreadyInitialized { .. })
