+++ crates/burn-backend/src/lib.rs
@@ -9,14 +9,6 @@ extern crate derive_new;
 
 extern crate alloc;
 
-mod data;
-pub use data::*;
-
-pub mod distribution;
-pub use distribution::*;
-pub mod element;
-pub use element::*;
-
 /// [`Backend`] trait and required types.
 pub mod backend;
 pub use backend::*;
@@ -24,8 +16,10 @@ pub use backend::*;
 // Re-exported types
 pub use burn_std::reader::*; // Useful so that backends don't have to add `burn_std` as a dependency.
 pub use burn_std::{
-    AllocationProperty, BoolDType, BoolStore, Bytes, DType, DeviceHandle, FloatDType, IntDType,
-    bf16, f16, stream_id::StreamId,
+    AllocationProperty, BoolDType, BoolStore, Bytes, DType, DataError, DeviceHandle, Distribution,
+    DistributionSampler, DistributionSamplerKind, Element, ElementConversion, ElementEq,
+    ElementOrdered, ElementRandom, FloatDType, IntDType, Scalar, TensorData, Tolerance, bf16,
+    distribution, element, f16, stream_id::StreamId,
 };
 
 /// Shape definition.
