--- crates/burn-backend/src/quantization/calibration.rs
@@ -1,5 +1,0 @@
-/// Calibration method used to compute the quantization range mapping.
-pub enum Calibration {
-    /// Computes quantization range mapping based on the min and max values.
-    MinMax,
-}
