+++ crates/burn-backend/src/quantization/mod.rs
@@ -1,12 +1,10 @@
-mod calibration;
 mod parameters;
 mod scheme;
 
-pub use calibration::*;
 pub use parameters::*;
 pub use scheme::*;
 
 pub use burn_std::quantization::{
-    BlockSize, QuantLevel, QuantMode, QuantParam, QuantPropagation, QuantScheme, QuantStore,
-    QuantValue, QuantizedBytes,
+    BlockSize, Calibration, QuantLevel, QuantMode, QuantParam, QuantPropagation, QuantScheme,
+    QuantStore, QuantValue, QuantizedBytes,
 };
