+++ crates/burn-std/src/data/mod.rs
