+++ crates/burn-std/src/distribution.rs
