+++ crates/burn-tensor/src/lib.rs
@@ -18,12 +18,18 @@ pub(crate) use tensor::check::macros::check;
 pub use tensor::*;
 
 // Re-exported types
-pub use burn_backend::{AllocationProperty, Bytes, StreamId, bf16, f16, read_sync, try_read_sync};
+pub use burn_std::{
+    AllocationProperty, Bytes, bf16, f16,
+    reader::{read_sync, try_read_sync},
+    stream_id::StreamId,
+};
 
 mod device;
 pub use device::*;
 
 pub(crate) use burn_backend::{
-    AutodiffBackend, Backend, BackendTypes, ExecutionError, element, get_device_settings,
-    tensor::{BoolTensor, FloatTensor, IntTensor},
+    AutodiffBackend, Backend, BackendTypes, ExecutionError, TensorMetadata, TensorPrimitive,
+    element, get_device_settings,
+    ops::TransactionPrimitive,
+    tensor::{BoolTensor, IntTensor},
 };
