+++ crates/burn-tensor/src/tensor/api/mod.rs
@@ -51,4 +51,3 @@ pub use float::{DEFAULT_ATOL, DEFAULT_RTOL};
 pub use options::*;
 pub use transaction::*;
 
-pub use burn_backend::tensor::IndexingUpdateOp;
