+++ crates/burn-tensor/src/tensor/mod.rs
@@ -5,10 +5,9 @@ mod api;
 pub use api::*;
 
 // Re-exported types
-pub use burn_backend::{
-    BoolDType, BoolStore, DType, DataError, FloatDType, IntDType, TensorData, TensorMetadata,
-    TensorPrimitive, Tolerance, distribution::*, element::*, indexing::*,
-    ops::TransactionPrimitive, shape::*, slice::*,
+pub use burn_std::{
+    BoolDType, BoolStore, DType, DataError, FloatDType, IndexingUpdateOp, IntDType, TensorData,
+    Tolerance, distribution::*, element::*, indexing::*, s, shape::*, slice::*,
 };
 
 /// The tensor kind module.
@@ -18,20 +17,19 @@ pub use kind::{Bool, Float, Int, TensorKind};
 /// The activation module.
 pub mod activation;
 
-/// The backend module.
+#[cfg(feature = "extension")]
+/// The backend module — escape hatch for backend extension authors.
+///
+/// Most users should not need this. It re-exports the backend implementer API
+/// (`Backend` trait, ops traits, tensor type aliases) under one path.
 pub mod backend {
-    #[cfg(not(feature = "extension"))]
-    pub use burn_backend::backend::ExecutionError;
-    #[cfg(feature = "distributed")]
-    pub use burn_backend::distributed;
-
-    #[cfg(feature = "extension")]
     pub use burn_backend::backend::*;
+    pub use burn_backend::tensor::{
+        BoolElem, BoolTensor, Device, FloatElem, FloatTensor, IntElem, IntTensor, QuantizedTensor,
+    };
 
-    #[cfg(feature = "extension")]
     pub use crate::bridge::{BasicOps, Numeric, Ordered};
 
-    #[cfg(feature = "extension")]
     /// The backend extension module.
     pub mod extension {
         pub use burn_backend_extension::backend_extension;
@@ -61,17 +59,7 @@ pub mod signal;
 
 /// Operations on tensors module.
 pub mod ops {
-    // Re-export explicit types without exposing the backend-level traits
-    // TODO: group them in burn-backend module instead, and re-export ::*
-    // pub use burn_backend::backend::ops::{
-    //     AttentionModuleOptions, ConvOptions, ConvTransposeOptions, DeformConvOptions,
-    //     GridSampleOptions, GridSamplePaddingMode, InterpolateMode, InterpolateOptions, PadMode,
-    //     PaddedConvOptions, TransactionPrimitive, UnfoldOptions,
-    // };
-    pub use burn_backend::backend::ops::*;
-    pub use burn_backend::tensor::{
-        BoolElem, BoolTensor, Device, FloatElem, FloatTensor, IntElem, IntTensor, QuantizedTensor,
-    };
+    pub use burn_std::ops::*;
 }
 
 /// Tensor quantization module.
