+++ crates/burn-tensor/src/tensor/quantization.rs
@@ -1,8 +1,8 @@
 use crate::{Tensor, TensorPrimitive};
 use burn_backend::quantization;
 
-// We re-export those types.
-pub use burn_backend::quantization::*;
+// User-facing quantization data types come from burn-std.
+pub use burn_std::quantization::*;
 use burn_dispatch::Dispatch;
 
 /// The tensor quantization parameters.
