+++ crates/burn-vision/src/backends/cube/connected_components/hardware_accelerated.rs
@@ -6,7 +6,11 @@
 use crate::{
     ConnectedStatsOptions, Connectivity, backends::cube::connected_components::stats_from_opts,
 };
-use burn_core::tensor::{Shape, TensorMetadata, cast::ToElement, ops::IntTensorOps};
+use burn_core::tensor::{
+    Shape,
+    backend::{TensorMetadata, ops::IntTensorOps},
+    cast::ToElement,
+};
 use burn_cubecl::{
     BoolElement, CubeBackend, CubeRuntime, FloatElement, IntElement, kernel,
     ops::{into_data_sync, numeric::zeros_client},
