+++ crates/burn-vision/src/backends/cube/connected_components/prefix_sum.rs
@@ -1,4 +1,4 @@
-use burn_core::tensor::{Shape, TensorMetadata};
+use burn_core::tensor::{Shape, backend::TensorMetadata};
 use cubecl::prelude::*;
 
 use burn_cubecl::{
