+++ examples/custom-wgpu-kernel/src/lib.rs
@@ -3,8 +3,8 @@ mod forward;
 
 use burn::backend::{autodiff::Autodiff, wgpu::Wgpu};
 use burn::tensor::backend::extension::{Dispatch, backend_extension};
-use burn::tensor::ops::FloatTensor; // TODO: should be moved to backend
-use burn::tensor::{Tensor, TensorPrimitive, activation};
+use burn::tensor::backend::{FloatTensor, TensorPrimitive};
+use burn::tensor::{Tensor, activation};
 
 /// We create our own Backend trait that extends the Burn backend trait.
 #[backend_extension(Autodiff, Wgpu)]
